fix size_t type in slice

+++ src/gausskernel/storage/access/pggate/slice.cpp
@@ -32,6 +32,7 @@
 // under the License.
 //
 
+#include <cstddef>
 #include <sstream>
 #include "slice.h"
 
@@ -81,7 +82,7 @@ std::string Slice::ToDebugString(size_t max_len) const {
     abbreviated = true;
   }
 
-  int num_not_graph = 0;
+  size_t num_not_graph = 0;
   for (size_t i = 0; i < bytes_to_print; i++) {
     if (!isgraph(begin_[i])) {
       ++num_not_graph;
@@ -95,7 +96,7 @@ std::string Slice::ToDebugString(size_t max_len) const {
   }
 
   std::ostringstream oss;
-  for (int i = 0; i < bytes_to_print; i++) {
+  for (size_t i = 0; i < bytes_to_print; i++) {
     auto ch = begin_[i];
     if (!isgraph(ch)) {
       if (ch == '\r') {
